requirements.txt
streamlit==1.41.1
python-dotenv
openai
langchain
langchain-community
